Pre-Phase 3: dashboard decomposition, strategy registry, immutable optimization, test coverage
Split 946-line dashboard/app.py into focused modules (charts, layouts,
callbacks, serialization, middleware) reducing it to 85 lines. Add
decorator-based strategy registry with data_requirement filtering.
Replace save/restore mutation pattern with Strategy.copy() for
thread-safe optimization. Push test coverage from 88% to 92%.

Closes #29, closes #30, closes #31, closes #45
