+++ strategies/__init__.py
@@ -1,4 +1,5 @@
 from strategies.base import Strategy
+from strategies.registry import registry, register, DataRequirement
 from strategies.moving_average_crossover import MovingAverageCrossover
 from strategies.rsi_strategy import RSIStrategy
 from strategies.bollinger_bands import BollingerBandsStrategy
